pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-stack"
version = "0.1.0"
description = "Open-source quantitative trading workflow — data to execution"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Core numerical computing
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",

    # Visualisation
    "matplotlib>=3.7",
    "seaborn>=0.12",

    # Config & utilities
    "pyyaml>=6.0",
    "pyarrow>=12.0",  # Parquet support for Pandas

    # ML - Classical
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
# Phase 2-3: AutoML and deep learning
ml-extended = [
    "pycaret>=3.0",
    "torch>=2.0",
]

# Phase 4: Portfolio optimisation & risk
portfolio = [
    "riskfolio-lib>=4.0",
    "alphalens-reloaded>=0.4",
    "pyfolio-reloaded>=0.9",
]

# Phase 5: Backtesting
backtest = [
    "vectorbt>=0.26",
]

# Phase 1 (network required): Live data sources
data-live = [
    "yfinance>=0.2",
    "openbb>=4.0",
]

# Phase 6: Live execution
execution = [
    "ibapi>=9.81",
]

# Research environment
research = [
    "jupyter>=1.0",
    "jupyterlab>=4.0",
    "ipywidgets>=8.0",
]

# Development tools
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
]

# Install everything
all = [
    "quant-stack[ml-extended,portfolio,backtest,data-live,execution,research,dev]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests requiring network access or external services",
    "slow: tests that take more than 10 seconds",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
